pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ccc"
version = "0.2.0"
description = "Claude Code Chat Bot for Telegram and Lark"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Herpiko Dwi Aguno"}
]
keywords = ["telegram", "lark", "bot", "claude", "ai", "development"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development",
]
dependencies = [
    "python-telegram-bot>=22.0",
    "PyYAML>=6.0",
    "claude-agent-sdk>=0.1.0",
    "lark-oapi>=1.0.0",
    "flask>=3.0.0",
    "pycryptodome>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.0",
    "mypy>=1.0",
]
telegram = [
    "python-telegram-bot>=22.0",
]
lark = [
    "lark-oapi>=1.0.0",
    "flask>=3.0.0",
    "pycryptodome>=3.0.0",
]

[project.scripts]
ccc = "ccc.__main__:main"
tgcc = "ccc.__main__:main"

[project.urls]
Homepage = "https://github.com/herpiko/tg-cc"
Repository = "https://github.com/herpiko/tg-cc"

[tool.setuptools.packages.find]
where = ["."]
include = ["ccc*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
